+++ pyfecons/inputs/power_input.py
@@ -10,13 +10,10 @@ class PowerInput:
     mn: Ratio = None  # Neutron energy multiplier
     eta_p: Percent = None  # Pumping power capture efficiency
     eta_th: Percent = None  # Thermal conversion efficiency
-    fpcppf: Percent = None  # Primary Coolant Pumping Power Fraction
     p_trit: MW = None  # Tritium Systems
     p_house: MW = None  # Housekeeping power
-    p_tfcool: MW = None  # Solenoid coil cooling
-    p_pfcool: MW = None  # Mirror coil cooling
-    p_tf: MW = None  # Power into TF coils
-    p_pf: MW = None  # Power into PF (equilibrium) coils (TODO - how to handle for HTS?)
+    p_cool: MW = None  # Coil cooling power (TF + PF)
+    p_coils: MW = None  # Power into coils (TF + PF)
     eta_pin: Percent = None  # Input power wall plug efficiency
     eta_pin1: Percent = None
     eta_pin2: Percent = None
@@ -25,4 +22,13 @@ class PowerInput:
     p_implosion: MW = None  # Implosion laser power
     p_ignition: MW = None  # Ignition laser power
     p_target: MW = None  # Power into target factory
-    p_machinery: MW = None  # Power into machinery
+    p_pump: MW = None  # Primary coolant pumping power
+    f_dec: Percent = (
+        None  # DEC charged-particle capture fraction (0-1). Default 0 (no DEC).
+    )
+    dd_f_T: Percent = None  # DD tritium burn fraction (0-1). Default 0.969.
+    dd_f_He3: Percent = None  # DD He-3 burn fraction (0-1). Default 0.689.
+    dhe3_dd_frac: Percent = (
+        None  # D-He3 energy fraction from D-D sides (0-1). Default 0.07.
+    )
+    dhe3_f_T: Percent = None  # D-He3 tritium burn in D-D sides (0-1). Default 0.97.